chore: create dateStore
