+++ store/index.ts
@@ -1,9 +1,11 @@
 import * as mapStore from './map';
 import * as dailyStore from './daily';
+import * as dateStore from './date';
 import * as weeklyStore from './weekly';
 
 const store = {
   map: mapStore,
+  date: dateStore,
   daily: dailyStore,
   weekly: weeklyStore,
 };
